mem: Add a parameter which will make a memory truly a ROM.
This piggy-backs on the writeOK method which already exists. It also
modifies the flags returned as part of the memory's backdoor
descriptor which doesn't enforce that the memory is read only, but will
let the other party know it's expected not to write to it.

Bug: 254411221
Test: Used this flag with the GSA's boot ROM and saw that it no longer
Test: accepted a corrupting write which overwrote the reset vector.
Change-Id: Ib95e619c76c327d302e62a88515a92af11815981
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/10032
Reviewed-by: Gabe Black <gabeblack@google.com>
Presubmit-Ready: Gabe Black <gabeblack@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Gem5-Virtual-Platform-Presubmit-Ready: Gabe Black <gabeblack@google.com>
Gem5-Virtual-Platform-Verified: kokoro <noreply+kokoro@google.com>