util: Add a missing free() to m5term.
The "cwd" string is used to store the previous working directory, if
m5term needs to change it to maximize the space it has to create a
unix domain socket using a relative path. That string was not being
freed, which is a small memory leak. This change frees that memory.

Change-Id: I1ad901e7c59614dd938165cae91c3f666d86e872
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/10373
Presubmit-Ready: Gabe Black <gabeblack@google.com>
Gem5-Virtual-Platform-Presubmit-Ready: Gabe Black <gabeblack@google.com>
Reviewed-by: Gabe Black <gabeblack@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>