base: fatal() if a socket path doesn't fit in sockaddr_un.sun_path.
Normally this would just generate a warning, but a warning is easy to
miss, and truncating the path to fit would be surprising. Since the max
length isn't likely to change, a path which has to be truncated is
essentially fundementally wrong, and could be defined as something
else which is short enough before being used in the config.

Note that this only applies to either the abstract path which is just
a string, or the file name and not the directory path on a file based
socket.

Bug: 277283635
Change-Id: I8702cf02c03053b5d0b6133f25b0e588de666f15
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/10492
Tested-by: kokoro <noreply+kokoro@google.com>
Gem5-Virtual-Platform-Presubmit-Ready: Gabe Black <gabeblack@google.com>
Presubmit-Ready: Gabe Black <gabeblack@google.com>
Reviewed-by: Gabe Black <gabeblack@google.com>